% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree.R
\name{calc_specific_agreement}
\alias{calc_specific_agreement}
\title{Calculate specific agreement}
\usage{
calc_specific_agreement(mat, category = 1)
}
\arguments{
\item{mat}{A matrix of coding where each row is an object of coding (i.e., an
action unit when doing description coding or a temporal bin when doing
location coding) and each column is a source of coding (e.g., human coder).}

\item{category}{A scalar (e.g., number, string, or boolean) that represents
which of the possible coding categories agreement should be calculated on.
Defaults to \code{1} for positive agreement during occurrence coding.}
}
\value{
A scalar double that is the specific agreement estimate from 0 to 1.
}
\description{
Calculate category-specific agreement using a generalized formula that can
accommodate missing data and any number of coders. With two raters, the
interpretation of specific agreement for any category is the probability of
one rater assigning an item to that category given that the other rater has
also assigned that item to that category. With more than two raters, the
interpretation becomes the probability of a randomly chosen rater assigning
an item to that category given that another randomly chosen rater has also
assigned that item to that category.
}
\details{
\deqn{A_k=\frac{\sum_{i=1}^{n'}r_{ik}(r_{ik}-1)}{\sum_{i=1}^{n'}r_{ik}(r_i-1)}}
\eqn{n'} is the number of objects that were assigned to any category by
two or more coders\if{html}{\out{<br>}}
\eqn{r_{ik}} is the number of coders that assigned object \eqn{i} to category
\eqn{k}\if{html}{\out{<br>}}
\eqn{r_i} is the number of coders that assigned object \eqn{i} to any category
}
