% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_occurrence}
\alias{compare_occurrence}
\title{Compare two codings on occurrence}
\usage{
compare_occurrence(x, y)
}
\arguments{
\item{x, y}{A pair of coding objects created by \code{\link[=coding]{coding()}}.}
}
\value{
A numeric vector containing the agreement scores between each
corresponding element of \code{x} and \code{y}.
}
\description{
Compare two FACS codings on which AUs are present vs. absent. Uses the Wexler
formula recommended in the 2002 FACS Investigator's Guide.
}
\details{
Let \eqn{P} be the number of AUs that occur in both the \code{x} and \code{y} codings
and let \eqn{Q} be the total number of AUs that occur in \code{x} and \code{y}
(counting repeated AUs twice). Wexler's formula is simply \eqn{2P/Q}.
}
\examples{
compare_occurrence(coding("1+4+9"), coding("1+4+10"))
}
