% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_proto.R
\name{compare_to_prototypes}
\alias{compare_to_prototypes}
\title{Compare FACS Coding Objects to Emotion Prototypes}
\usage{
compare_to_prototypes(
  x,
  scheme,
  proto = prototypes,
  uncoded = c("warn", "error", "omit"),
  output = c("all", "exact", "closest")
)
}
\arguments{
\item{x}{An object created by \code{coding()} to contain one or more FACS codings.}

\item{scheme}{An object created by \code{facs_scheme()} to indicate which AUs were
coded and in which ways (i.e., occurrence, intensity, asymmetry).}

\item{proto}{A FUNCTION TO CREATE PROTOTYPES NEEDS TO BE MADE.}

\item{uncoded}{A string indicating how to handle the situation where a
prototype contains an AU that wasn't included in the scheme. "warn"
provides a warning and an NA value, "error" provides an error and stops,
and "omit" drops the uncoded AU before calculating the Dice coefficient.}

\item{output}{A string indicating what to return. "all" returns the Dice
coefficients for all prototypes, "exact" returns only the prototypes where
the Dice coefficient equals 1.0, and "closest" returns only the prototypes
with the maximum observed Dice coefficient.}
}
\value{
A list the same length as \code{x} wherein each element contains a named
vector containing zero or more prototypes and Dice coefficients quantifying
their match to the coding in \code{x}.
}
\description{
How similar is this coding to each emotion prototype?
}
