% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{facs_codes}
\alias{facs_codes}
\title{List of Official FACS Codes}
\format{
\subsection{\code{facs_codes}}{

A data frame with 73 rows and 5 columns:
\describe{
\item{type}{Code type}
\item{number}{Code number}
\item{name}{Code name}
\item{group}{Code group}
\item{chapter}{Primary chapter from FACS 2002 manual}
}
}
}
\source{
Ekman, P., Friesen, W. V., & Hager, J. (2002). Facial action coding
system: A technique for the measurement of facial movement. Research Nexus.
}
\usage{
facs_codes
}
\description{
Information about each of the official FACS codes including action units
(AUs), action descriptors (ADs), and supplementary codes (SCs).
}
\keyword{datasets}
