\name{findpeaks1}
\alias{findpeaks1}
\title{Find peaks function}
\usage{
findPeaks1 <- function(file,tolerance=.995,peaks=5, resol=250,plot=F)}
\arguments{
\item{file}{a dna file}
\item{tolerance}{limit for the peak detection (need to be more specific...)}
\item{peaks}{number of desired peaks to find}
\item{resol}{resolution of the kernel, default to 250}
\item{plot}{default TRUE}
}
\description{
Add something here
}

