% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prototypes}
\alias{prototypes}
\title{List of all available emotion prototypes}
\format{
\subsection{\code{prototypes}}{

A data frame with 606 rows and 7 columns:
\describe{
\item{source}{A string describing the source of the prototype.}
\item{emotion}{A string describing the name of the emotion.}
\item{emotion_type}{An optional string describing the emotion type.}
\item{config_num}{A numerical ID within each source-emotion combo.}
\item{config_type}{A string describing the configuration type.}
\item{code}{FACS coding string describing each configuration.}
\item{special}{A string with additional non-FACS information.}
}
}
}
\usage{
prototypes
}
\description{
Description
}
\keyword{datasets}
