% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scheme_full}
\alias{scheme_full}
\title{Full coding scheme will all FACS codes}
\format{
An object created by \code{\link[=scheme]{scheme()}}
}
\usage{
scheme_full
}
\description{
For user convenience, \code{scheme_full} is provided that contains occurrence,
intensity, and asymmetry coding for all codes possible in FACS 2002.
}
\keyword{datasets}
